+++ core/src/test/java/com/syntaric/openfhir/mapping/eps/EpsToAqlTest.java
@@ -1,0 +1,211 @@
+package com.syntaric.openfhir.mapping.eps;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.syntaric.openfhir.OpenFhirMappingContext;
+import com.syntaric.openfhir.TestOpenFhirMappingContext;
+import com.syntaric.openfhir.aql.ToAqlRequest;
+import com.syntaric.openfhir.aql.ToAqlResponse;
+import com.syntaric.openfhir.aql.ToAqlResponse.AqlResponse;
+import com.syntaric.openfhir.manager.FhirConnectManager;
+import com.syntaric.openfhir.manager.OptManager;
+import com.syntaric.openfhir.db.entity.FhirConnectContextEntity;
+import com.syntaric.openfhir.db.entity.OptEntity;
+import com.syntaric.openfhir.fc.schema.context.FhirConnectContext;
+import com.syntaric.openfhir.mapping.helpers.AqlToFlatPathConverter;
+import com.syntaric.openfhir.mapping.helpers.HelpersCreator;
+import com.syntaric.openfhir.mapping.toaql.OpenEhrAqlPopulator;
+import com.syntaric.openfhir.mapping.toaql.ToAql;
+import com.syntaric.openfhir.mapping.toaql.ToAqlMappingEngine;
+import com.syntaric.openfhir.util.*;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FileUtils;
+import org.ehrbase.openehr.sdk.webtemplate.parser.OPTParser;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.openehr.schemas.v1.OPERATIONALTEMPLATE;
+import org.openehr.schemas.v1.TemplateDocument;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doReturn;
+
+@Slf4j
+public class EpsToAqlTest {
+    AutoCloseable closeable;
+
+    @Mock
+    protected FhirConnectManager fhirConnectManager;
+
+    @Mock
+    protected OptManager optManager;
+
+    @Mock
+    protected OpenEhrTemplateUtils openEhrTemplateUtils;
+
+    final String DIR = getClass().getResource("/eps/").getFile();
+    final String CONTEXT_DIR = getClass().getResource("/eps/").getFile();
+
+    private ToAql toAql;
+
+    @Before
+    public void setupState() {
+        closeable = MockitoAnnotations.openMocks(this);
+
+        final TestOpenFhirMappingContext repo = new TestOpenFhirMappingContext(new FhirConnectModelMerger());
+
+        final List<FhirConnectContextEntity> allContextEntities = contextFiles().stream().map(x -> {
+            FhirConnectContextEntity fhirConnectContextEntity = new FhirConnectContextEntity();
+            fhirConnectContextEntity.setFhirConnectContext(parseContext(x));
+            return fhirConnectContextEntity;
+        }).toList();
+        doReturn(allContextEntities).when(fhirConnectManager).allUserContextEntities();
+
+        for (String contextFile : contextFiles()) {
+            try {
+                final FhirConnectContext context = parseContext(contextFile);
+                final OPERATIONALTEMPLATE operationalTemplate = getOperationalTemplate();
+                repo.initRepository(context, operationalTemplate, DIR);
+
+                final FhirConnectContextEntity toBeReturned = new FhirConnectContextEntity();
+                toBeReturned.setFhirConnectContext(context);
+
+                final String templateId = operationalTemplate.getTemplateId().getValue();
+                final String normalizedTemplateId = OpenFhirMappingContext.normalizeTemplateId(templateId);
+                doReturn(toBeReturned).when(fhirConnectManager).findContextByTemplateId(eq(normalizedTemplateId));
+
+                final OptEntity optEntity = new OptEntity();
+                optEntity.setContent(getOperationalTemplateContent());
+                doReturn(new OPTParser(operationalTemplate).parse()).when(openEhrTemplateUtils).parseWebTemplate(eq(optEntity));
+                doReturn(optEntity).when(optManager).byTemplateIdAndOrganization(eq(normalizedTemplateId));
+
+            } catch (Exception e) {
+                log.error("{}", e.getMessage());
+            }
+        }
+
+        final HelpersCreator helpersCreator1 = new HelpersCreator(repo, new AqlToFlatPathConverter(
+                new OpenFhirStringUtils(),
+                new OpenFhirMapperUtils()), new OpenFhirStringUtils());
+        toAql = new ToAql(fhirConnectManager, new OpenFhirMapperUtils(), repo, new ToAqlMappingEngine(new OpenEhrAqlPopulator()), helpersCreator1,
+                openEhrTemplateUtils, null, optManager);
+    }
+
+    @Test
+    public void toAql_allergyIntolerance() {
+        final ToAqlResponse response = toAql.toAql(new ToAqlRequest(null, "123", "AllergyIntolerance"));
+        assertAql(response, ToAqlResponse.AqlType.ENTRY,
+                "SELECT h FROM EHR e CONTAINS EVALUATION h [openEHR-EHR-EVALUATION.adverse_reaction_risk.v2] WHERE e/ehr_id/value='{{ehrid}}'");
+        assertAql(response, ToAqlResponse.AqlType.COMPOSITION,
+                "SELECT c FROM EHR e CONTAINS COMPOSITION c CONTAINS EVALUATION h [openEHR-EHR-EVALUATION.adverse_reaction_risk.v2] WHERE e/ehr_id/value='{{ehrid}}'");
+    }
+
+    @Test
+    public void toAql_conditionVerificationStatusConfirmed() {
+        final ToAqlResponse response = toAql.toAql(new ToAqlRequest(null, "123", "Condition?verification-status=confirmed"));
+        assertAql(response, ToAqlResponse.AqlType.ENTRY,
+                "SELECT h FROM EHR e CONTAINS EVALUATION h [openEHR-EHR-EVALUATION.problem_diagnosis.v1] WHERE e/ehr_id/value='{{ehrid}}' AND h/data[at0001]/items[at0073]/value/defining_code/terminology_id/value = 'local' AND h/data[at0001]/items[at0073]/value/defining_code/code_string = 'at0076' AND h/data[at0001]/items[at0073]/value/value = 'Confirmed' AND h/data[at0001]/items[at0073]/value/defining_code/terminology_id/value = 'local' AND h/data[at0001]/items[at0073]/value/defining_code/code_string = 'at0076' AND h/data[at0001]/items[at0073]/value/value = 'Confirmed'");
+        assertAql(response, ToAqlResponse.AqlType.COMPOSITION,
+                "SELECT c from EHR e CONTAINS COMPOSITION c WHERE e/ehr_id/value='{{ehrid}}' and c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Problems']/items[openEHR-EHR-EVALUATION.problem_diagnosis.v1]/data[at0001]/items[at0073]/value/defining_code/terminology_id/value = 'local' AND c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Problems']/items[openEHR-EHR-EVALUATION.problem_diagnosis.v1]/data[at0001]/items[at0073]/value/defining_code/code_string = 'at0076' AND c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Problems']/items[openEHR-EHR-EVALUATION.problem_diagnosis.v1]/data[at0001]/items[at0073]/value/value = 'Confirmed' AND c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Problems']/items[openEHR-EHR-EVALUATION.problem_diagnosis.v1]/data[at0001]/items[at0073]/value/defining_code/terminology_id/value = 'local' AND c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Problems']/items[openEHR-EHR-EVALUATION.problem_diagnosis.v1]/data[at0001]/items[at0073]/value/defining_code/code_string = 'at0076' AND c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Problems']/items[openEHR-EHR-EVALUATION.problem_diagnosis.v1]/data[at0001]/items[at0073]/value/value = 'Confirmed'");
+    }
+
+    @Test
+    public void toAql_procedure() {
+        final ToAqlResponse response = toAql.toAql(new ToAqlRequest(null, "123", "Procedure"));
+        assertAql(response, ToAqlResponse.AqlType.ENTRY,
+                "SELECT h FROM EHR e CONTAINS ACTION h [openEHR-EHR-ACTION.procedure.v1] WHERE e/ehr_id/value='{{ehrid}}'");
+        assertAql(response, ToAqlResponse.AqlType.COMPOSITION,
+                "SELECT c FROM EHR e CONTAINS COMPOSITION c CONTAINS ACTION h [openEHR-EHR-ACTION.procedure.v1] WHERE e/ehr_id/value='{{ehrid}}'");
+    }
+
+    @Test
+    public void toAql_deviceUseStatement() {
+        final ToAqlResponse response = toAql.toAql(new ToAqlRequest(null, "123", "DeviceUseStatement?_include=DeviceUseStatement:device"));
+        assertAql(response, ToAqlResponse.AqlType.ENTRY,
+                "SELECT h FROM EHR e CONTAINS EVALUATION h [openEHR-EHR-EVALUATION.device_summary.v0] WHERE e/ehr_id/value='{{ehrid}}'");
+        assertAql(response, ToAqlResponse.AqlType.COMPOSITION,
+                "SELECT c FROM EHR e CONTAINS COMPOSITION c CONTAINS EVALUATION h [openEHR-EHR-EVALUATION.device_summary.v0] WHERE e/ehr_id/value='{{ehrid}}'");
+    }
+
+    @Test
+    public void toAql_byCodedText() {
+        final ToAqlResponse response = toAql.toAql(new ToAqlRequest(null, "123", "AllergyIntolerance?manifestation=4386001"));
+        assertAql(response, ToAqlResponse.AqlType.ENTRY,
+                "SELECT h FROM EHR e CONTAINS CLUSTER h [openEHR-EHR-EVALUATION.adverse_reaction_risk.v2] WHERE e/ehr_id/value='{{ehrid}}' AND h/data[at0001]/items[openEHR-EHR-CLUSTER.adverse_reaction_event.v1]/items[at0006]/value/value = '4386001'");
+        assertAql(response, ToAqlResponse.AqlType.COMPOSITION,
+                "SELECT c from EHR e CONTAINS COMPOSITION c WHERE e/ehr_id/value='{{ehrid}}' and c/content[openEHR-EHR-SECTION.adhoc.v1 and name/value='EPS Allergies']/items[openEHR-EHR-EVALUATION.adverse_reaction_risk.v2]/data[at0001]/items[openEHR-EHR-CLUSTER.adverse_reaction_event.v1]/items[at0006]/value/value = '4386001'");
+    }
+
+    private void assertAql(final ToAqlResponse response, final ToAqlResponse.AqlType type, final String expectedAql) {
+        Assert.assertNotNull(response);
+        final AqlResponse match = response.getAqls().stream()
+                .filter(a -> a.getType() == type)
+                .findFirst()
+                .orElse(null);
+        Assert.assertNotNull("No AQL of type " + type + " found in response", match);
+        Assert.assertEquals(expectedAql, match.getAql());
+    }
+
+    private OPERATIONALTEMPLATE getOperationalTemplate() {
+        try {
+            return TemplateDocument.Factory.parse(getOperationalTemplateContent()).getTemplate();
+        } catch (final Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private String getOperationalTemplateContent() {
+        final File optDir = new File(DIR, "");
+        final File[] optFiles = optDir.listFiles(f -> f.getName().endsWith(".opt"));
+        if (optFiles == null || optFiles.length == 0) {
+            throw new RuntimeException("No .opt file found in " + optDir.getAbsolutePath());
+        }
+        try {
+            return FileUtils.readFileToString(optFiles[0]);
+        } catch (final Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private FhirConnectContext parseContext(String path) {
+        final ObjectMapper yaml = OpenFhirTestUtility.getYaml();
+        try {
+            return yaml.readValue(getFileContent(path), FhirConnectContext.class);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @SneakyThrows
+    private String getFileContent(final String filePath) {
+        return FileUtils.readFileToString(new File(filePath), StandardCharsets.UTF_8);
+    }
+
+    private List<String> contextFiles() {
+        final List<String> result = new ArrayList<>();
+        collectContextFiles(new File(CONTEXT_DIR), result);
+        return result;
+    }
+
+    private void collectContextFiles(final File dir, final List<String> result) {
+        final File[] files = dir.listFiles();
+        if (files == null) {
+            return;
+        }
+        for (final File file : files) {
+            if (file.isDirectory()) {
+                collectContextFiles(file, result);
+            } else if (file.getName().endsWith("context.yml")) {
+                result.add(file.getAbsolutePath());
+            }
+        }
+    }
+}